Add explicit transactions (#60)
This adds the concepts of transactions to the client. A transaction works at a given Noms commit and a MapEditor is created so that reads and writes are returning consistent values. If the transaction does any writes and Commit is called, the MapEditor is materialized and the Noms Map is committed (FastForward).

This is exposed to the SDK clients as openTransaction which returns a transaction ID which needs to be passed along to all operations.

Towards rocicorp/replicache#27